SPI Ethernet Refactoring
